Handle pemasukan
