--- Pemasukan.py
@@ -1,94 +1,0 @@
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-class Ui_Pemasukan(object):
-    def setupUi(self, Dialog):
-        Dialog.setObjectName("Dialog")
-        Dialog.resize(360, 300)
-        Dialog.setMinimumSize(QtCore.QSize(360, 300))
-        Dialog.setMaximumSize(QtCore.QSize(360, 300))
-        self.label = QtWidgets.QLabel(Dialog)
-        self.label.setGeometry(QtCore.QRect(40, 10, 221, 51))
-        font = QtGui.QFont()
-        font.setFamily("Montserrat ExtraBold")
-        font.setPointSize(14)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label.setFont(font)
-        self.label.setStyleSheet("color: rgb(57, 171, 121);")
-        self.label.setObjectName("label")
-        self.dateEdit = QtWidgets.QDateEdit(Dialog)
-        self.dateEdit.setGeometry(QtCore.QRect(130, 70, 191, 31))
-        self.dateEdit.setDateTime(QtCore.QDateTime(QtCore.QDate(2020, 1, 1), QtCore.QTime(0, 0, 0)))
-        self.dateEdit.setObjectName("dateEdit")
-        self.label_4 = QtWidgets.QLabel(Dialog)
-        self.label_4.setGeometry(QtCore.QRect(40, 71, 81, 31))
-        font = QtGui.QFont()
-        font.setFamily("Montserrat SemiBold")
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setItalic(False)
-        font.setWeight(7)
-        self.label_4.setFont(font)
-        self.label_4.setStyleSheet("font: 63 10pt \"Montserrat SemiBold\";")
-        self.label_4.setObjectName("label_4")
-        self.label_5 = QtWidgets.QLabel(Dialog)
-        self.label_5.setGeometry(QtCore.QRect(40, 121, 81, 31))
-        font = QtGui.QFont()
-        font.setFamily("Montserrat SemiBold")
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setItalic(False)
-        font.setWeight(7)
-        self.label_5.setFont(font)
-        self.label_5.setStyleSheet("font: 63 10pt \"Montserrat SemiBold\";")
-        self.label_5.setObjectName("label_5")
-        self.label_6 = QtWidgets.QLabel(Dialog)
-        self.label_6.setGeometry(QtCore.QRect(40, 171, 81, 31))
-        font = QtGui.QFont()
-        font.setFamily("Montserrat SemiBold")
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setItalic(False)
-        font.setWeight(7)
-        self.label_6.setFont(font)
-        self.label_6.setStyleSheet("font: 63 10pt \"Montserrat SemiBold\";")
-        self.label_6.setObjectName("label_6")
-        self.pushButton = QtWidgets.QPushButton(Dialog)
-        self.pushButton.setGeometry(QtCore.QRect(40, 221, 281, 41))
-        self.pushButton.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.pushButton.setMouseTracking(False)
-        self.pushButton.setStyleSheet("border-radius: 5px;\n"
-"font: 57 10pt \"Montserrat Medium\";\n"
-"background-color: rgb(57, 171, 121);\n"
-"color: rgb(255, 255, 255);")
-        self.pushButton.setObjectName("pushButton")
-        self.lineEdit = QtWidgets.QLineEdit(Dialog)
-        self.lineEdit.setGeometry(QtCore.QRect(130, 170, 191, 31))
-        self.lineEdit.setObjectName("lineEdit")
-        self.lineEdit_2 = QtWidgets.QLineEdit(Dialog)
-        self.lineEdit_2.setGeometry(QtCore.QRect(130, 120, 191, 31))
-        self.lineEdit_2.setObjectName("lineEdit_2")
-
-        self.retranslateUi(Dialog)
-        QtCore.QMetaObject.connectSlotsByName(Dialog)
-
-    def retranslateUi(self, Dialog):
-        _translate = QtCore.QCoreApplication.translate
-        Dialog.setWindowTitle(_translate("Dialog", "Dialog"))
-        self.label.setText(_translate("Dialog", "Tambah Pemasukan"))
-        self.label_4.setText(_translate("Dialog", "Tanggal"))
-        self.label_5.setText(_translate("Dialog", "Nominal"))
-        self.label_6.setText(_translate("Dialog", "Keterangan"))
-        self.pushButton.setText(_translate("Dialog", "Tambah"))
-        self.lineEdit.setPlaceholderText(_translate("Dialog", "Masukkan sumber pemasukan"))
-        self.lineEdit_2.setPlaceholderText(_translate("Dialog", "Masukkan nominal pemasukan"))
-
-
-if __name__ == "__main__":
-    import sys
-    app = QtWidgets.QApplication(sys.argv)
-    Dialog = QtWidgets.QDialog()
-    ui = Ui_Pemasukan()
-    ui.setupUi(Dialog)
-    Dialog.show()
-    sys.exit(app.exec_())
