--- Riwayat.py
@@ -1,108 +1,0 @@
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-class Ui_Riwayat(object):
-    def setupUi(self, Dialog):
-        Dialog.setObjectName("Dialog")
-        Dialog.resize(500, 500)
-        Dialog.setMinimumSize(QtCore.QSize(500, 500))
-        Dialog.setMaximumSize(QtCore.QSize(500, 500))
-        self.label = QtWidgets.QLabel(Dialog)
-        self.label.setGeometry(QtCore.QRect(40, 10, 101, 51))
-        font = QtGui.QFont()
-        font.setFamily("Montserrat ExtraBold")
-        font.setPointSize(14)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label.setFont(font)
-        self.label.setStyleSheet("color: rgb(57, 171, 121);")
-        self.label.setObjectName("label")
-        self.tableWidget = QtWidgets.QTableWidget(Dialog)
-        self.tableWidget.setGeometry(QtCore.QRect(40, 61, 417, 400))
-        self.tableWidget.setObjectName("tableWidget")
-        self.tableWidget.setColumnCount(4)
-        self.tableWidget.setRowCount(8)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setVerticalHeaderItem(0, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setVerticalHeaderItem(1, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setVerticalHeaderItem(2, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setVerticalHeaderItem(3, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setVerticalHeaderItem(4, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setVerticalHeaderItem(5, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setVerticalHeaderItem(6, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setVerticalHeaderItem(7, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setHorizontalHeaderItem(0, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setHorizontalHeaderItem(1, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setHorizontalHeaderItem(2, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setHorizontalHeaderItem(3, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setItem(0, 0, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setItem(0, 1, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setItem(0, 2, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setItem(0, 3, item)
-
-        self.retranslateUi(Dialog)
-        QtCore.QMetaObject.connectSlotsByName(Dialog)
-
-    def retranslateUi(self, Dialog):
-        _translate = QtCore.QCoreApplication.translate
-        Dialog.setWindowTitle(_translate("Dialog", "Dialog"))
-        self.label.setText(_translate("Dialog", "Riwayat"))
-        item = self.tableWidget.verticalHeaderItem(0)
-        item.setText(_translate("Dialog", "1"))
-        item = self.tableWidget.verticalHeaderItem(1)
-        item.setText(_translate("Dialog", "2"))
-        item = self.tableWidget.verticalHeaderItem(2)
-        item.setText(_translate("Dialog", "3"))
-        item = self.tableWidget.verticalHeaderItem(3)
-        item.setText(_translate("Dialog", "4"))
-        item = self.tableWidget.verticalHeaderItem(4)
-        item.setText(_translate("Dialog", "5"))
-        item = self.tableWidget.verticalHeaderItem(5)
-        item.setText(_translate("Dialog", "6"))
-        item = self.tableWidget.verticalHeaderItem(6)
-        item.setText(_translate("Dialog", "8"))
-        item = self.tableWidget.verticalHeaderItem(7)
-        item.setText(_translate("Dialog", "9"))
-        item = self.tableWidget.horizontalHeaderItem(0)
-        item.setText(_translate("Dialog", "Tanggal"))
-        item = self.tableWidget.horizontalHeaderItem(1)
-        item.setText(_translate("Dialog", "Aksi"))
-        item = self.tableWidget.horizontalHeaderItem(2)
-        item.setText(_translate("Dialog", "Nominal"))
-        item = self.tableWidget.horizontalHeaderItem(3)
-        item.setText(_translate("Dialog", "Keterangan"))
-        __sortingEnabled = self.tableWidget.isSortingEnabled()
-        self.tableWidget.setSortingEnabled(False)
-        item = self.tableWidget.item(0, 0)
-        item.setText(_translate("Dialog", "12/12/2020"))
-        item = self.tableWidget.item(0, 1)
-        item.setText(_translate("Dialog", "Pemasukan"))
-        item = self.tableWidget.item(0, 2)
-        item.setText(_translate("Dialog", "30000"))
-        item = self.tableWidget.item(0, 3)
-        item.setText(_translate("Dialog", "Gaji"))
-        self.tableWidget.setSortingEnabled(__sortingEnabled)
-
-
-if __name__ == "__main__":
-    import sys
-    app = QtWidgets.QApplication(sys.argv)
-    Dialog = QtWidgets.QDialog()
-    ui = Ui_Riwayat()
-    ui.setupUi(Dialog)
-    Dialog.show()
-    sys.exit(app.exec_())
